+++ src/app/meals/create/recipe-section.tsx
@@ -1,12 +1,19 @@
 'use client';
 
+import styles from '@/styles/app/meals/create/recipe-form.module.scss';
 import { MealRecipeSectionWithId } from '@/src/types/meal.types';
 import { InputString } from '@/src/components/common/form/input-string';
 import { useState } from 'react';
 import { useRecipeFormContext } from '@/src/contexts/recipe-form.context';
-import { createNewRecipeStep, updateSectionName, updateSectionStep } from '@/src/helpers/recipe-form.helper';
-import { Button } from '@/src/components/common/button';
+import {
+    createNewRecipeStep,
+    removeSection,
+    updateSectionName,
+    updateSectionStep
+} from '@/src/helpers/recipe-form.helper';
 import { RecipeSectionStep } from '@/src/app/meals/create/recipe-section-step';
+import { AddButton } from '@/src/components/common/add-button';
+import { RemoveButton } from '@/src/components/common/remove-button';
 
 interface RecipeSectionProps {
     section: MealRecipeSectionWithId;
@@ -30,13 +37,22 @@ export function RecipeSection({ section }: RecipeSectionProps) {
         onChangeSections(modifiedSections);
     };
 
+    const onRemoveSection = () => {
+        const modifiedSections = removeSection(section.id, sections);
+
+        onChangeSections(modifiedSections);
+    };
+
     return (
         <li>
-            <InputString label={'Section name'} value={name} setValue={setValue} />
-            <ul>
+            <div className={styles['section-name-container']}>
+                <InputString label={'Section name'} value={name} setValue={setValue} />
+                <RemoveButton label={'Remove'} onClick={onRemoveSection} />
+            </div>
+            <ol>
                 {section.steps.map(step => <RecipeSectionStep key={step.id} step={step} section={section} />)}
-            </ul>
-            <Button label={'Add recipe step'} onClick={onAddStep} />
+            </ol>
+            <AddButton label={'Add the next step'} onClick={onAddStep} />
         </li>
     );
 }