+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.32.0] - 2024-08-22
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `AddButton` component as combination of icon and text.
+- [Marek Kurańda](https://github.com/mjkuranda): `RemoveButton` component as combination of icon and text.
+- [Marek Kurańda](https://github.com/mjkuranda): The ability to remove sections and section steps within a recipe.
+
+### Changed
+- [Marek Kurańda](https://github.com/mjkuranda): Substituted adding button at meal creation page for `AddButton`.
+
 ## [1.31.1] - 2024-08-22
 ### Changed
 - [Marek Kurańda](https://github.com/mjkuranda): At least 2 steps for the first recipe section are required.